backend/requirements.txt
requests==2.26.0
asgiref==3.2.10
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
django-filter==21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.20.0
djoser==2.1.0
pillow==9.2.0